requirements.txt
streamlit
streamlit_option_menu
pandas
requests
numpy
cognite-sdk
python-dotenv